hash in push pop

+++ src/stack_ctor_dtor.cpp
@@ -29,7 +29,7 @@ stack_result_t stack_ctor(struct stack *stk, int capacity, const char *file_name
     *data = canary;
     stk->data = data + canary_shift;
 
-    for(size_t i = 0; i < stk->capacity; i++)
+    for(int i = 0; i < stk->capacity; i++)
     {
         stk->data[i] = poison;
     }
